requirements.txt
alive-progress==3.1.4
pyyaml==6.0